[Feat] 페이스정보 보여주기 #10

+++ TrackUs-iOS/TrackUs-iOS/Model/Running.swift
@@ -14,7 +14,7 @@ struct Running {
     var calorie: Double = 0.0
     var distance: Double = 0.0
     var pace: Double = 0.0
-    var seconds: Int = 0
+    var seconds: Double = 0.0
     var geoPoints: [GeoPoint] = []
     var isGroup: Bool = false
     var routeImageUrl: String = ""
